update-relocate-MainPage-to-main-page dir
