+++ src/components/main-page/Footer.vue
