+++ src/components/main-page/RectWideBtn.vue
