+++ src/components/main-page/Sectioin2.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { getData } from '../apiFunction'
+import { getData } from '../../apiFunction'
 
 const token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2N2Y0YjA4YTU4ZWNkOGYzNWY4N2Y2YWUiLCJpYXQiOjE3NDQwODkyMzEsImV4cCI6MTc0NDY5NDAzMX0.-XGmTM0X7ptHggl4nnyL1YR3-l0Tr2tsJlJurz4z3kg"
 const newsUrl = "https://team-rocket-hotelapi-from-freyja.onrender.com/api/v1/admin/news"
