+++ src/components/main-page/Section6.vue
