+++ src/components/main-page/Slider.vue
